+++ noodles-bcf/CHANGELOG.md
@@ -14,6 +14,9 @@
 
     Use `Info::iter` for fields.
 
+  * bcf/record/info: Change `Info::get` to return an `Option<Value>` value
+    instead of a field.
+
 ## 0.19.2 - 2022-12-02
 
 ### Fixed